+++ playground/vanilla/app/controllers/health_controller.rb
@@ -1,0 +1,6 @@
+# :nodoc:
+class HealthController < ApplicationController
+  def show
+    head :ok
+  end
+end